+++ Structurizr.Core/View/RankDirection.cs
@@ -1,0 +1,11 @@
+namespace Structurizr
+{
+    public enum RankDirection
+    {
+        TopBottom,
+        BottomTop,
+        LeftRight,
+        RightLeft
+    }
+    
+}